+++ src/encoded/types/__init__.py
@@ -108,46 +108,6 @@ class Enzyme(Item):
 
 
 @collection(
-    name='biosources',
-    unique_key='accession',
-    properties={
-        'title': 'Biosources',
-        'description': 'Cell lines and tissues used for biosamples',
-    })
-class Biosource(Item):
-    """Biosource class."""
-
-    item_type = 'biosource'
-    name_key = 'accession'
-    schema = load_schema('encoded:schemas/biosource.json')
-    embedded = ["individual", "individual.organism"]
-
-    @calculated_property(schema={
-        "title": "Biosource name",
-        "description": "Specific name of the biosource.",
-        "type": "string",
-    })
-    def biosource_name(self, request, biosource_type, individual=None, cell_line=None, tissue=None):
-        if biosource_type == "tissue":
-            if tissue:
-                return tissue
-        elif biosource_type == "immortalized cell line":
-            if cell_line:
-                return cell_line
-        elif biosource_type == "primary cell":
-            if cell_line:
-                return cell_line
-        elif biosource_type == "whole organisms":
-            if individual:
-                individual_props = request.embed(individual, '@@object')
-                organism = individual_props['organism']
-                organism_props = request.embed(organism, '@@object')
-                organism_name = organism_props['name']
-                return "whole " + organism_name
-        return biosource_type
-
-
-@collection(
     name='constructs',
     properties={
         'title': 'Constructs',
